move remote data to @api directory
