+++ template/.prettierrc.js
@@ -14,7 +14,6 @@ module.exports = {
     '^@hoc/(.*)$',
     '^@hooks/(.*)$',
     '^@localization/(.*)$',
-    '^@models/(.*)$',
     '^@navigation/(.*)$',
     '^@redux/(.*)$',
     '^@screens/(.*)$',
