+++ template/src/api/authSlice.ts
@@ -2,10 +2,10 @@ import { PayloadAction, createSlice } from '@reduxjs/toolkit'
 import { REHYDRATE, persistReducer } from 'redux-persist'
 import { PersistPartial } from 'redux-persist/es/persistReducer'
 import { call, put, takeLatest, takeLeading } from 'typed-redux-saga'
+import { Failure, Loading, NotRequested, RemoteData, Success, isSuccess } from '@api/RemoteData'
 import { logIn as logInRequest } from '@api/auth'
 import { AuthTokens, setAuthConfig } from '@api/common'
 import { Credentials } from '@api/types/auth.types'
-import { Failure, Loading, NotRequested, RemoteData, Success, isSuccess } from '@models/RemoteData'
 import { safeStorage } from '@redux/persistence'
 import { resetStore } from '@redux/rootActions'
 import { RootState } from '@redux/store'
