+++ template/src/screens/DemoScreen.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { ActivityIndicator, Button, Image, StyleSheet, Text } from 'react-native'
+import { hasData, isNotRequested } from '@api/RemoteData'
 import MainScreenLayout from '@components/layouts/MainScreenLayout'
 import DemoCard from '@components/surfaces/DemoCard'
 import { TestIDs } from '@config/testIDs'
 import Colors from '@config/ui/colors'
 import useAppDispatch from '@hooks/useAppDispatch'
 import useAppSelector from '@hooks/useAppSelector'
-import { hasData, isNotRequested } from '@models/RemoteData'
 import type { RootStackScreenProps } from '@navigation/navigators/RootStackNavigator'
 import Routes from '@navigation/routes'
 import { clearPersistence } from '@redux/persistence'
