+++ template/src/screens/LoginScreen.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { ActivityIndicator, Button, StyleSheet } from 'react-native'
+import { isLoading } from '@api/RemoteData'
 import { logInAsync, selectAuthTokens } from '@api/authSlice'
 import DemoTextInput from '@components/inputs/DemoTextInput'
 import MainScreenLayout from '@components/layouts/MainScreenLayout'
 import DemoCard from '@components/surfaces/DemoCard'
 import useAppDispatch from '@hooks/useAppDispatch'
 import useAppSelector from '@hooks/useAppSelector'
-import { isLoading } from '@models/RemoteData'
 import { RootStackScreenProps } from '@navigation/navigators/RootStackNavigator'
 import Routes from '@navigation/routes'
