+++ template/tsconfig.json
@@ -46,7 +46,6 @@
       "@hoc/*": ["./hoc/*"],
       "@hooks/*": ["./hooks/*"],
       "@localization/*": ["./localization/*"],
-      "@models/*": ["./models/*"],
       "@navigation/*": ["./navigation/*"],
       "@redux/*": ["./redux/*"],
       "@screens/*": ["./screens/*"],
